chore: extract and test
