+++ package.json
@@ -26,7 +26,7 @@
     "@graphql-tools/merge": "^8.3.3",
     "fp-async-utils": "^2.0.1",
     "glob": "^8.0.3",
-    "graphql": "^16.4.0"
+    "graphql": "^16.5.0"
   },
   "devDependencies": {
     "@zeit/ncc": "^0.22.3",
@@ -35,7 +35,7 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^26.1.5",
     "eslint-plugin-prettier": "4.0.0",
-    "jest": "^28.0.3",
+    "jest": "^28.1.3",
     "prettier": "2.6.2"
   }
 }
