+++ src/index.js
@@ -1,87 +1,14 @@
 const fs = require('fs').promises;
 const { dirname } = require('path');
 const core = require('@actions/core');
-const { gqlPluckFromCodeString } = require('@graphql-tools/graphql-tag-pluck');
-const { mergeTypeDefs } = require('@graphql-tools/merge');
-const { glob } = require('glob');
-const { asyncPipe, asyncMap, asyncFilter } = require('fp-async-utils');
-
-/**
- * @param {string} source
- * @returns {Promise<string[]>}
- */
-async function getFilepaths(source) {
-  return new Promise((resolve, reject) => {
-    glob(source, { nonull: true }, (err, filePaths) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(filePaths);
-      }
-    });
-  });
-}
-
-/**
- * @param {string} filePath
- * @returns {Promise<{filePath: string, content: string}>}
- */
-async function getContent(filePath) {
-  const content = await fs.readFile(filePath, 'utf8');
-  return { filePath, content };
-}
-
-/**
- *
- * @param {{filePath: string, content: string}} param0
- * @returns {Promise<string>}
- */
-async function pluckGQL({ filePath, content }) {
-  const [plucked] = await gqlPluckFromCodeString(filePath, content);
-  return plucked && plucked.body;
-}
-
-/**
- * @param {string[]} schemas
- * @returns {Promise<string>}
- */
-async function mergeGql(schemas) {
-  return mergeTypeDefs(schemas);
-}
-
-/**
- *
- * @param {DocumentNode} schemaDocumentNode
- * @returns {Promise<string>}
- */
-async function extractSchemaString(schemaDocumentNode) {
-  return schemaDocumentNode.loc.source.body;
-}
-
-/**
- *
- * @param {string} schema
- * @returns {Promise<void>}
- */
-async function writeSchemaToOutput(schema) {
-  const output = core.getInput('output');
-  core.info(`Writing to file ${output}`);
-  await fs.writeFile(output, schema);
-}
+const pluckSchema = require('./pluck-schema');
 
 async function main() {
-  await asyncPipe(
-    core.getInput('source'),
-    getFilepaths,
-    asyncMap(getContent),
-    asyncMap(pluckGQL),
-    asyncFilter(Boolean),
-    mergeGql,
-    extractSchemaString,
-    writeSchemaToOutput
-  );
+  const source = core.getInput('source');
+  const schema = await pluckSchema(source);
 
   const output = core.getInput('output');
+  await fs.writeFile(core.getInput('output'), schema);
 
   core.debug(`Setting filepath to ${output}`);
   core.setOutput('filepath', output);
