+++ test/schema/schema1.js
@@ -1,0 +1,11 @@
+var gql;
+
+gql`
+  type TheInput1 {
+    someField: String!
+  }
+
+  type Query {
+    query1(input: TheInput1): String
+  }
+`;
