+++ test/schema/schema2.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-unused-vars
+var noSchemaInHere = true;
