+++ test/schema/schema3.js
@@ -1,0 +1,11 @@
+var gql;
+
+gql`
+  type TheInput3 {
+    someField: String!
+  }
+
+  type Query {
+    query3(input: TheInput3): String
+  }
+`;
